+++ gulpfile.js/index.js
@@ -13,13 +13,13 @@ var
 	// tasks
 	styles = require('./tasks/styles'),
 	stylesMin = require('./tasks/styles_min'),
+	stylesRTL = require('./tasks/styles_rtl')
 	scripts = require('./tasks/scripts'),
 	scriptsMin = require('./tasks/scripts_min'),
 	images = require('./tasks/images'),
 	imagesMin = require('./tasks/images_min'),
 	templates = require('./tasks/templates'),
 	sprites = require('./tasks/sprites/sprites'),
-	autoprefixer = require('./tasks/autoprefixer'),
 	watch = require('./tasks/watch'),
 	bower = require('./tasks/bower'),
 	browsersync = require('./tasks/browsersync'),
@@ -32,6 +32,7 @@ var
 // ##################################################################################
 gulp.task('styles', styles);
 gulp.task('styles:min', stylesMin);
+gulp.task('styles:rtl', stylesRTL);
 
 gulp.task('scripts', scripts);
 gulp.task('scripts:min', scriptsMin);
@@ -45,8 +46,6 @@ gulp.task('sprites', sprites);
 
 gulp.task('bower', bower);
 
-gulp.task('autoprefixer', autoprefixer);
-
 gulp.task('watch', watch);
 
 gulp.task('browsersync', browsersync);
@@ -57,17 +56,11 @@ gulp.task('browsersync', browsersync);
 
 // Run all tasks
 gulp.task('default',
-	gulp.series(
-		gulp.parallel('bower', 'images', 'styles', 'scripts', 'templates', 'sprites'),
-		gulp.series('autoprefixer')
-	)
+	gulp.series('bower', 'images', 'styles', 'scripts', 'templates', 'sprites')
 );
 
 gulp.task('dev',
-	gulp.parallel(
-		gulp.series('default', 'browsersync'),
-		'watch'
-	)
+	gulp.series('default', 'browsersync', 'watch')
 );
 
 gulp.task('minify',