+++ gulpfile.js/tasks/styles_rtl.js
@@ -1,19 +1,17 @@
 var
 	gulp = require('gulp'),
 	$ = require('gulp-load-plugins')(),
-	autoprefixer = require('autoprefixer'),
 
 	config = require('../config'),
 	consoleError = require('../utils/console_error');
 
-function prefixer() {
+function stylesRTL() {
 	return gulp.src( config.paths.built.styles.all )
-		.pipe(
-			$.postcss(
-				[ autoprefixer() ]
-			)
-		)
-		.pipe( gulp.dest( config.paths.built.styles.path ) );
+		.pipe( $.rtlcss() )
+		.pipe( $.rename({
+			suffix: '-rtl'
+		}) )
+		.pipe( gulp.dest(config.paths.built.styles.path) )
 };
 
-module.exports = prefixer;
+module.exports = stylesRTL;